test(light): 携行光源の転写をテストで固定し冗長キャストを削る
StatsChangedSystem の光源転写に単体テストが無く、クエリ変更等で静かに壊れうる。
トーチ装備→点灯・解除→消灯の両方向を固定するテストを足す。あわせて既に float64 の
info.Darkness への冗長な float64 キャストを削る。

+++ internal/systems/vision.go
@@ -106,7 +106,7 @@ func (sys *VisionSystem) Update(world w.World) error {
 		info := calculateLightSourceDarkness(world, consts.Coord[int]{X: tileData.Col, Y: tileData.Row}, blockViewIndex, ambient)
 		// 明るさが閾値未満なら見えない。視界を光の届く範囲へ寄せる。
 		// 見えないタイルは記憶側へ回るので、暗所の敵やアイテムは自然に隠れる
-		if 1.0-float64(info.Darkness) < visibilityThreshold {
+		if 1.0-info.Darkness < visibilityThreshold {
 			continue
 		}
 		vs.LightSourceCache[gridElement] = info
